requirements.txt
aiogram==3.2.0
elevenlabs==0.2.26
langchain==0.0.342
openai==1.3.5
pydantic==2.5.2
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0